+++ montreal_forced_aligner/command_line/g2p.py
@@ -8,6 +8,7 @@ from praatio import tgio
 from montreal_forced_aligner.multiprocessing.corpus import parse_transcription
 from montreal_forced_aligner.corpus.align_corpus import AlignableCorpus
 from montreal_forced_aligner.helper import load_text
+from montreal_forced_aligner.config.g2p_config import load_basic_g2p_config, g2p_yaml_to_config
 
 from montreal_forced_aligner.g2p.generator import PyniniDictionaryGenerator as Generator
 from montreal_forced_aligner.models import G2PModel
@@ -18,7 +19,7 @@ from montreal_forced_aligner.exceptions import ArgumentError
 from montreal_forced_aligner.config import TEMP_DIR
 
 
-def generate_dictionary(args):
+def generate_dictionary(args, unknown_args=None):
     print("Generating pronunciations from G2P model")
     if not args.temp_directory:
         temp_dir = TEMP_DIR
@@ -28,6 +29,12 @@ def generate_dictionary(args):
     if args.clean:
         shutil.rmtree(os.path.join(temp_dir, 'G2P'), ignore_errors=True)
         shutil.rmtree(os.path.join(temp_dir, 'models', 'G2P'), ignore_errors=True)
+    if args.config_path:
+        g2p_config = g2p_yaml_to_config(args.config_path)
+    else:
+        g2p_config = load_basic_g2p_config()
+    if unknown_args:
+        g2p_config.update_from_args(unknown_args)
     if os.path.isdir(args.input_path):
         input_dir = os.path.expanduser(args.input_path)
         corpus_name = os.path.basename(args.input_path)
@@ -36,7 +43,8 @@ def generate_dictionary(args):
             corpus_name = os.path.basename(args.input_path)
         data_directory = os.path.join(temp_dir, corpus_name)
 
-        corpus = AlignableCorpus(input_dir, data_directory, num_jobs=args.num_jobs, use_mp=(not args.disable_mp))
+        corpus = AlignableCorpus(input_dir, data_directory, num_jobs=args.num_jobs, use_mp=(not args.disable_mp),
+                                 punctuation=g2p_config.punctuation, clitic_markers=g2p_config.clitic_markers)
 
         word_set = get_word_set(corpus, args.include_bracketed)
     else:
@@ -51,7 +59,7 @@ def generate_dictionary(args):
         model = G2PModel(args.g2p_model_path, root_directory=os.path.join(temp_dir, 'models', 'G2P'))
         model.validate(word_set)
         gen = Generator(model, word_set, temp_directory=temp_dir, num_jobs=args.num_jobs,
-                        num_pronunciations=args.num_pronunciations)
+                        num_pronunciations=g2p_config.num_pronunciations)
         gen.output(args.output_path)
         model.clean_up()
     else:
@@ -121,17 +129,17 @@ def validate(args, pretrained_languages):
         raise (ArgumentError('Could not find the input path {}.'.format(args.input_path)))
 
 
-def run_g2p(args, pretrained=None):
+def run_g2p(args, unknown=None, pretrained=None):
     if pretrained is None:
         pretrained = get_available_g2p_languages()
     validate(args, pretrained)
-    generate_dictionary(args)
+    generate_dictionary(args, unknown)
 
 
 if __name__ == '__main__':  # pragma: no cover
     from montreal_forced_aligner.command_line.mfa import g2p_parser, fix_path, unfix_path, g2p_languages
-    g2p_args = g2p_parser.parse_args()
+    g2p_args, unknown_args = g2p_parser.parse_known_args()
 
     fix_path()
-    run_g2p(g2p_args, g2p_languages)
+    run_g2p(g2p_args, unknown_args, g2p_languages)
     unfix_path()