+++ montreal_forced_aligner/corpus/base.py
@@ -16,12 +16,16 @@ supported_audio_extensions = ['.flac', '.ogg', '.aiff']
 def get_wav_info(file_path, sample_rate=16000):
     with soundfile.SoundFile(file_path, 'r') as inf:
         subtype = inf.subtype
-        bit_depth = int(subtype.split('_')[-1])
+        if subtype == 'FLOAT':
+            bit_depth = 32
+        else:
+            bit_depth = int(subtype.split('_')[-1])
         frames = inf.frames
         sr = inf.samplerate
         duration = frames / sr
         return_dict = {'num_channels': inf.channels, 'type': inf.subtype, 'bit_depth': bit_depth,
                        'sample_rate': sr, 'duration': duration, 'format': inf.format}
+    print(return_dict)
     use_sox = False
     if bit_depth != 16:
         use_sox = True